Use Typescript instead of Babel
Replace scaffolded files with Typescript version instead of Babel
version.  This will make development easier and hopefully more
bug-free.
