--- mean-stack/yeoman-mean/client/.eslintrc
@@ -1,13 +1,0 @@
-{
-    "extends": "../.eslintrc",
-    "env": {
-        "browser": true,
-        "commonjs": true
-    },
-    "globals": {
-        "process": true
-    },
-    "rules": {
-        "no-process-env": 0
-    }
-}
