+++ mean-stack/yeoman-mean/client/app/app.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app',
-  template: `<navbar></navbar>
+    selector: 'app',
+    template: `<navbar></navbar>
     <router-outlet></router-outlet>
     <footer></footer>`
 })
