--- mean-stack/yeoman-mean/client/app/app.constants.js
@@ -1,2 +1,0 @@
-export default from '../../server/config/environment/shared';
-
