+++ mean-stack/yeoman-mean/client/app/app.constants.ts
@@ -1,0 +1,4 @@
+
+import shared from '../../server/config/environment/shared.js';
+
+export default shared;
