+++ mean-stack/yeoman-mean/client/app/polyfills.ts
@@ -2,9 +2,12 @@ import 'core-js/es6';
 import 'core-js/es7/reflect';
 import 'zone.js/dist/zone';
 
+interface IPolyFillErrorConstructor extends ErrorConstructor {
+    stackTraceLimit: any;
+}
 
 if(!ENV) {
-  var ENV = 'development';
+    var ENV = 'development';
 }
 
 if(ENV === 'production') {
@@ -12,7 +15,6 @@ if(ENV === 'production') {
 } else {
     // Development
 
-  Error.stackTraceLimit = Infinity;
-
+    (<IPolyFillErrorConstructor>Error).stackTraceLimit = Infinity;
     // require('zone.js/dist/long-stack-trace-zone');
 }
