--- mean-stack/yeoman-mean/client/components/footer/footer.component.js
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'footer',
-  template: require('./footer.html'),
-  styles: [require('./footer.scss')]
-})
-export class FooterComponent {}
