+++ mean-stack/yeoman-mean/client/components/footer/footer.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'footer',
+    template: require('./footer.html'),
+    styles: [require('./footer.scss')]
+})
+export class FooterComponent {}
