--- mean-stack/yeoman-mean/client/components/oauth-buttons/oauth-buttons.component.js
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'oauth-buttons',
-  template: require('./oauth-buttons.html'),
-  styles: [require('./oauth-buttons.scss')],
-})
-export class OauthButtonsComponent {
-  loginOauth(provider) {
-    window.location.href = `/auth/${provider}`;
-  }
-}
