+++ mean-stack/yeoman-mean/client/components/socket/primus.mock.ts
