+++ mean-stack/yeoman-mean/client/components/socket/socket.mock.ts
@@ -1,6 +1,6 @@
 'use strict';
 
 export class SocketServiceStub {
-  syncUpdates() {}
-  unsyncUpdates() {}
+    syncUpdates() {}
+    unsyncUpdates() {}
 }
