--- mean-stack/yeoman-mean/client/components/util.spec.js
@@ -1,10 +1,0 @@
-import { safeCb } from './util';
-import { expect } from 'chai';
-
-describe('Util', () => {
-  it('Has a safeCb function', () => {
-    let notAFunction;
-
-    expect(safeCb(notAFunction)).to.not.throw(Error);
-  });
-});
