+++ mean-stack/yeoman-mean/client/components/util.spec.ts
@@ -1,0 +1,10 @@
+import { safeCb } from './util';
+import { expect } from 'chai';
+
+describe('Util', () => {
+    it('Has a safeCb function', () => {
+        let notAFunction = undefined;
+
+        expect(safeCb(notAFunction)).to.not.throw(Error);
+    });
+});
