+++ mean-stack/yeoman-mean/package.json
@@ -18,6 +18,7 @@
     "lodash": "^4.17.4",
     "lodash-es": "^4.17.4",
     "lusca": "^1.4.1",
+    "babel-core": "^6.18.2",
     "babel-plugin-angular2-annotations": "^5.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.6.0",
@@ -44,6 +45,10 @@
     "serve-favicon": "^2.3.0"
   },
   "devDependencies": {
+    "@angularclass/hmr-loader": "^3.0.4",
+    "reflect-metadata": "^0.1.3",
+    "rxjs": "^5.5.2",
+    "zone.js": "^0.8.12",
     "@angular/common": "^5.0.1",
     "@angular/compiler": "^5.0.1",
     "@angular/core": "^5.0.1",
@@ -53,40 +58,38 @@
     "@angular/platform-browser-dynamic": "^5.0.1",
     "@angular/router": "^5.0.1",
     "@angularclass/hmr": "^2.1.3",
-    "@angularclass/hmr-loader": "^3.0.4",
-    "angular-validation-match": "^1.9.0",
-    "angular2-jwt": "^0.2.3",
+    "@types/core-js": "^0.9.41",
+    "@types/lodash": "^4.14.63",
+    "@types/mocha": "^2.2.41",
+    "@types/chai": "^4.0.4",
+    "@types/sinon": "^2.1.3",
+    "@types/sinon-chai": "^2.7.27",
+    "@types/node": "^8.0.26",
+    "@types/selenium-webdriver": "^3.0.3",
+    "@types/webpack": "^3.0.10",
     "angular2-universal": "2.1.0-rc.1",
-    "autoprefixer": "^7.1.3",
-    "awesome-typescript-loader": "^3.3.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-istanbul": "^4.1.4",
-    "babel-plugin-syntax-flow": "^6.8.0",
-    "babel-plugin-transform-class-properties": "^6.6.0",
-    "babel-plugin-transform-flow-comments": "^6.8.0",
-    "babel-plugin-transform-runtime": "^6.6.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-register": "^6.16.0",
+    "angular2-jwt": "^0.2.3",
+    "angular-validation-match": "^1.9.0",
     "bootstrap": "~3.3.7",
+    "ngx-bootstrap": "~1.6.3",
     "bootstrap-sass": "~3.3.7",
     "bootstrap-social": "^5.1.1",
-    "bs-fullscreen-message": "^1.0.0",
-    "chai": "^4.1.2",
-    "chai-as-promised": "^7.1.1",
-    "chai-things": "^0.2.0",
-    "css-loader": "^0.28.7",
-    "del": "^3.0.0",
-    "eslint": "^2.12.0",
-    "extract-text-webpack-plugin": "3.0.0",
-    "file-loader": "^0.11.2",
-    "flow-bin": "^0.54.0",
     "font-awesome": ">=4.1.0",
-    "grunt": "^1.0.1",
-    "grunt-build-control": "^0.7.0",
+    "autoprefixer": "^7.1.3",
+    "babel-core": "^6.6.5",
+    "babel-eslint": "^7.2.3",
+    "babel-register": "^6.16.0",
+    "bs-fullscreen-message": "^1.0.0",
+    "babel-plugin-transform-class-properties": "^6.6.0",
+    "babel-plugin-transform-runtime": "^6.6.0",
+    "babel-plugin-istanbul": "^4.1.4",
+    "babel-preset-es2015": "^6.6.0",
+    "eslint": "^2.12.0",
+    "del": "^3.0.0",
     "gulp": "^3.9.1",
     "gulp-babel": "^7.0.0",
+    "gulp-typescript": "~3.2.2",
+    "gulp-tslint": "^8.0.0",
     "gulp-env": "^0.4.0",
     "gulp-eslint": "^2.0.0",
     "gulp-imagemin": "^3.0.1",
@@ -101,51 +104,58 @@
     "gulp-rev-replace": "^0.4.2",
     "gulp-sort": "^2.0.0",
     "gulp-sourcemaps": "^2.6.1",
-    "gulp-stylint": "^4.0.0",
     "gulp-util": "^3.0.5",
     "gulp-watch": "^4.3.5",
-    "html-webpack-harddisk-plugin": "~0.1.0",
-    "html-webpack-plugin": "^2.24.1",
-    "imports-loader": "^0.7.1",
+    "gulp-stylint": "^4.0.0",
+    "grunt": "^1.0.1",
+    "grunt-build-control": "^0.7.0",
     "isparta": "^4.0.0",
-    "istanbul": "1.1.0-alpha.1",
-    "istanbul-instrumenter-loader": "^3.0.0",
-    "karma": "^1.7.1",
-    "karma-chai-plugins": "~0.9.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.0.0",
-    "karma-firefox-launcher": "^1.0.0",
-    "karma-mocha": "^1.0.1",
-    "karma-phantomjs-launcher": "~1.0.0",
-    "karma-script-launcher": "^1.0.0",
-    "karma-sourcemap-loader": "~0.3.7",
-    "karma-spec-reporter": "~0.0.20",
-    "karma-webpack": "^2.0.3",
-    "lazypipe": "^1.0.1",
-    "mocha": "^3.0.2",
-    "ngx-bootstrap": "~1.6.3",
-    "node-sass": "^4.5.3",
     "nodemon": "^1.3.7",
-    "null-loader": "^0.1.1",
-    "open": "~0.0.4",
-    "phantomjs-prebuilt": "^2.1.4",
-    "postcss-loader": "^2.0.6",
-    "proxyquire": "^1.0.1",
-    "raw-loader": "^0.5.1",
-    "reflect-metadata": "^0.1.3",
     "run-sequence": "^2.1.0",
-    "rxjs": "^5.5.2",
-    "sass-loader": "^6.0.6",
-    "sinon": "^3.2.1",
-    "sinon-chai": "^2.8.0",
-    "strip-ansi": "^4.0.0",
-    "style-loader": "^0.18.2",
-    "supertest": "^3.0.0",
-    "through2": "^2.0.1",
-    "typescript": "~2.6.1",
+    "lazypipe": "^1.0.1",
     "webpack": "^3.5.5",
     "webpack-dev-server": "^2.4.2",
-    "zone.js": "^0.8.12"
+    "extract-text-webpack-plugin": "3.0.0",
+    "html-webpack-plugin": "^2.24.1",
+    "html-webpack-harddisk-plugin": "~0.1.0",
+    "typescript": "~2.6.1",
+    "awesome-typescript-loader": "^3.3.0",
+    "babel-loader": "^7.1.2",
+    "css-loader": "^0.28.7",
+    "file-loader": "^0.11.2",
+    "imports-loader": "^0.7.1",
+    "istanbul-instrumenter-loader": "^3.0.0",
+    "null-loader": "^0.1.1",
+    "postcss-loader": "^2.0.6",
+    "raw-loader": "^0.5.1",
+    "style-loader": "^0.18.2",
+    "node-sass": "^4.5.3",
+    "sass-loader": "^6.0.6",
+    "karma-webpack": "^2.0.3",
+    "through2": "^2.0.1",
+    "open": "~0.0.4",
+    "istanbul": "1.1.0-alpha.1",
+    "chai": "^4.1.2",
+    "sinon": "^3.2.1",
+    "chai-as-promised": "^7.1.1",
+    "chai-things": "^0.2.0",
+    "karma": "^1.7.1",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-script-launcher": "^1.0.0",
+    "karma-coverage": "^1.0.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-phantomjs-launcher": "~1.0.0",
+    "karma-spec-reporter": "~0.0.20",
+    "karma-sourcemap-loader": "~0.3.7",
+    "sinon-chai": "^2.8.0",
+    "mocha": "^3.0.2",
+    "karma-mocha": "^1.0.1",
+    "karma-chai-plugins": "~0.9.0",
+    "phantomjs-prebuilt": "^2.1.4",
+    "proxyquire": "^1.0.1",
+    "strip-ansi": "^4.0.0",
+    "supertest": "^3.0.0",
+    "tslint": "^5.7.0"
   },
   "engines": {
     "node": "^6.2.2",
@@ -153,11 +163,10 @@
   },
   "scripts": {
     "lint": "npm run lint:client && npm run lint:server",
-    "lint:client": "eslint ./client/**/*.js -c ./client/.eslintrc --ignore-pattern *.spec.js",
+    "lint:client": "eslint ./client/**/*.ts -c ./client/.eslintrc --ignore-pattern *.spec.ts",
     "lint:server": "eslint ./server/**/*.js -c ./server/.eslintrc --ignore-pattern *.spec.js --ignore-pattern *.integration.js",
     "test": "gulp test",
     "test:client": "karma start ./karma.conf.js --single-run",
-    "flow": "flow",
     "update-webdriver": "node node_modules/protractor/bin/webdriver-manager update",
     "start": "node server",
     "start:client": "webpack-dev-server --config webpack.dev.js",
