+++ mean-stack/yeoman-mean/spec.js
@@ -10,11 +10,11 @@ import 'zone.js/dist/mocha-patch';
 import 'zone.js/dist/async-test';
 import 'zone.js/dist/fake-async-test';
 
-var testsContext = require.context('./client', true, /\.(spec|test)\.js$/);
+var testsContext = require.context('./client', true, /\.(spec|test)\.ts$/);
 // testsContext.keys().forEach(testsContext);
-testsContext('./app/main/main.component.spec.js');
-testsContext('./components/util.spec.js');
-testsContext('./components/oauth-buttons/oauth-buttons.component.spec.js');
+testsContext('./app/main/main.component.spec.ts');
+testsContext('./components/util.spec.ts');
+testsContext('./components/oauth-buttons/oauth-buttons.component.spec.ts');
 
 import { TestBed, getTestBed } from '@angular/core/testing';
 import {
