+++ mean-stack/yeoman-mean/tsconfig.client.test.json
@@ -1,0 +1,16 @@
+{
+  "compilerOptions": {
+    "sourceMap": true,
+    "rootDir": "./client",
+    "module": "commonjs",
+    "outDir": ".tmp/test"
+  },
+  "moduleResolution": "node",
+  "typeRoots": [
+    "node_modules/@types"
+  ],
+  "filesGlob": [
+    "client/{app,components}/**/*.{spec,mock}.ts",
+    "client/test_typings/**/*.d.ts"
+  ]
+}
