+++ mean-stack/yeoman-mean/tsconfig.json
@@ -1,0 +1,37 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "allowSyntheticDefaultImports": true,
+    "baseUrl": ".",
+    "experimentalDecorators": true,
+    "sourceMap": true,
+    "rootDir": "./",
+    "module": "es6",
+    "outDir": ".tmp",
+    "paths": {
+      "primus": ["client/components/socket/primus.js"]
+    },
+    "removeComments": false,
+    "target": "es5",
+    "skipLibCheck": true,
+    "moduleResolution": "node",
+    "lib": [
+      "es2016",
+      "dom"
+    ]
+  },
+  "typeRoots": [
+    "./node_modules/@types/"
+  ],
+  "exclude": [
+    "node_modules",
+    "client/**/*.spec.ts"
+  ],
+  "filesGlob": [
+    "client/{app,components}/**/!(*.spec).ts"
+  ],
+  "awesomeTypescriptLoaderOptions": {
+    "resolveGlobs": true,
+    "forkChecker": true
+  }
+}
